requirements.txt
django==4.0.1
djongo==1.3.6
pymongo==3.12.3
python-dotenv
statsmodels
requests
pandas
google-api-python-client
scikit-learn